// apps/frontend/src/shared/ui/generated/field.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * Controlled Field data. Native DOM attributes, refs and callbacks are technical React bindings.
 */
export interface FieldConfiguration {
  label: string;
  value: string;
  size?: 's' | 'm' | 'l';
  helper?: string;
  error?: string;
  placeholder?: string;
  disabled?: boolean;
  readOnly?: boolean;
  clearable?: boolean;
  type?: 'text' | 'email' | 'tel' | 'url' | 'search' | 'password';
  options?: FieldOption[];
}
export interface FieldOption {
  value: string;
  label: string;
  disabled?: boolean;
}